added quick explination to top of tutorial
The original tutorial threw you to the wolves a little too quickly.  As
an engineer and not a programmer, it took me a while to understand what
was going on with pint as there are not a lot of usage examples on the
web to just copy as is my usual MO.  The concepts are written in a logical
order rather than alphabetical, this is the future, you can use ^f to find
anything specific.  I still need to figure out how to format the docs,
I have no idea what I am doing.